# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateRepairRequestInput:
    properties:
      description:
        example: Обнаружен люфт в подшипнике главного вала
        type: string
      equipment_id:
        example: 1
        type: integer
      priority:
        example: high
        type: string
      title:
        example: Замена подшипника
        type: string
    type: object
...
  handlers.UpdateRepairRequestInput:
    properties:
      assigned_to:
        example: 2
        type: integer
      status:
        example: in_progress
        type: string
    type: object
...
  models.Equipment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      installation_date:
        type: string
      inventory_number:
        type: string
      last_maintenance_date:
        type: string
      location:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      serial_number:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.MaintenanceSchedule:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      description:
        type: string
      equipment:
        $ref: '#/definitions/models.Equipment'
      equipment_id:
        type: integer
      id:
        type: integer
      interval_days:
        type: integer
      is_active:
        type: boolean
      last_performed:
        type: string
      next_date:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.RepairRequest:
    properties:
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.User'
      completed_at:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      description:
        type: string
      equipment:
        $ref: '#/definitions/models.Equipment'
      equipment_id:
        type: integer
      id:
        type: integer
      priority:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API веб-приложения для управления техническим обслуживанием и ремонтом
    оборудования предприятия
  title: ТОиР API
  version: "1.0"
paths:
...
  /maintenance-schedules:
    get:
      description: Получение списка расписаний технического обслуживания с пагинацией
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Элементов на странице
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MaintenanceSchedule'
                  type: array
                meta:
                  $ref: '#/definitions/response.Meta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Список расписаний ТО
      tags:
      - maintenance-schedules
    post:
      consumes:
      - application/json
      description: Создаёт новое расписание планового технического обслуживания
      parameters:
      - description: Данные расписания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MaintenanceSchedule'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MaintenanceSchedule'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание расписания ТО
      tags:
      - maintenance-schedules
  /maintenance-schedules/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет расписание планового технического обслуживания
      parameters:
      - description: ID расписания
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MaintenanceSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MaintenanceSchedule'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Обновление расписания ТО
      tags:
      - maintenance-schedules
  /repair-requests:
    get:
      description: Получение списка заявок с пагинацией и фильтрами
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Элементов на странице
        in: query
        name: per_page
        type: integer
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Фильтр по приоритету
        in: query
        name: priority
        type: string
      - description: Фильтр по исполнителю
        in: query
        name: assigned_to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RepairRequest'
                  type: array
                meta:
                  $ref: '#/definitions/response.Meta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Список заявок на ремонт
      tags:
      - repair-requests
    post:
      consumes:
      - application/json
      description: Создаёт новую заявку на ремонт оборудования
      parameters:
      - description: Данные заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRepairRequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RepairRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание заявки на ремонт
      tags:
      - repair-requests
  /repair-requests/{id}:
    get:
      description: Возвращает заявку на ремонт по идентификатору
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RepairRequest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение заявки по ID
      tags:
      - repair-requests
    put:
      consumes:
      - application/json
      description: Обновляет статус или назначает исполнителя заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRepairRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Обновление заявки на ремонт
      tags:
      - repair-requests
...
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"